+++ src/protoclustplsr.jl
@@ -5,7 +5,7 @@ struct ProtoClustPlsr
 end
 
 ## Not exported
-function protoclustplsr(X, y; metric = :eucl, nproto, nlv, kavg = 1, h = 1, criw = 3, squared = false, 
+function protoclustplsr(X, y; metric = :eucl, nproto, nlv, kavg = 1, h = 1, criw = 4, squared = false, 
         tolw = 1e-4, scal = false)
     if metric == :eucl
         distance = Distances.Euclidean()